+++ colossalai/shardformer/policies/autopolicy.py
@@ -56,6 +56,8 @@ _POLICY_LIST = {
         PolicyLocation(file_name="t5", class_name="T5EncoderPolicy"),
 
     # GPT2
+    "transformers.models.gpt2.modeling_gpt2.GPT2Model":
+        PolicyLocation(file_name="gpt2", class_name="GPT2ModelPolicy"),
 }
 
 
@@ -98,5 +100,4 @@ def get_autopolicy(model: nn.Module) -> Policy:
         )
     else:
         policy = import_policy(policy_location)
-    return policy()
     return policy()
